+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './components/App';
-import store from './store/store';
+import store from './utils/store';
 import './index.css';
 
 ReactDOM.render(
