+++ src/utils/spoonacular.js
@@ -8,8 +8,10 @@ import {
   changeType,
 } from '../actions/index';
 
+const key = 'e58f966470464622b23fe6ed8e79d89d'
+
 const recipesType = async type => {
-  const response = await fetch(`https://api.spoonacular.com/recipes/complexSearch?type=${type}&apiKey=f48b01151d62432cb1902a6cbd7c2e8d`);
+  const response = await fetch(`https://api.spoonacular.com/recipes/complexSearch?type=${type}&apiKey=${key}`);
 
   if (response.ok) return response.json();
 
@@ -17,7 +19,7 @@ const recipesType = async type => {
 };
 
 const recipeProps = async id => {
-  const response = await fetch(`https://api.spoonacular.com/recipes/${id}/information?includeNutrition=false&apiKey=f48b01151d62432cb1902a6cbd7c2e8d`);
+  const response = await fetch(`https://api.spoonacular.com/recipes/${id}/information?includeNutrition=false&apiKey=${key}`);
 
   if (response.ok) return response.json();
